# openapi.yaml
openapi: 3.0.0
info:
  title: 'Q Laravel Skeleton API'
  version: '0.9'
paths:
...
  '/api/v1/users/{user}':
    get:
      tags:
        - Users
      summary: 'Get user data'
      operationId: getUserV1
      parameters:
        -
          name: user
          in: path
          description: 'User id'
          required: true
          example: 1
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                properties:
                  user: { $ref: '#/components/schemas/UserResource' }
                type: object
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthorized
        '404':
          description: 'Not found'
      security:
        -
          sanctum: []
    put:
      tags:
        - Users
      summary: 'Update existing user'
      operationId: updateUserV1
      parameters:
        -
          name: user
          in: path
          description: 'User id'
          required: true
          example: 1
      requestBody:
        description: 'Update existing user'
        required: true
        content:
          application/json:
            schema:
              required:
                - first_name
                - last_name
                - email
                - password
              properties:
                first_name:
                  type: string
                  example: John
                last_name:
                  type: string
                  example: Doe
                email:
                  type: string
                  example: john.doe@q.agency
                password:
                  type: string
                  example: secretPassword
              type: object
      responses:
        '200':
          description: 'User updated successfully'
          content:
            application/json:
              schema:
                properties:
                  user: { $ref: '#/components/schemas/UserResource' }
                type: object
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthorized
        '422':
          description: 'Validation failed'
        '404':
          description: 'Not found'
      security:
        -
          sanctum: []
    delete:
      tags:
        - Users
      summary: 'Delete user'
      operationId: deleteUserV1
      parameters:
        -
          name: user
          in: path
          description: 'User id'
          required: true
          example: 1
      responses:
        '204':
          description: Deleted
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'User deleted.' }
                type: object
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthorized
        '404':
          description: 'Not found'
      security:
        -
          sanctum: []
...
  '/api/v2/users/{user}':
    get:
      tags:
        - Users
      summary: 'Get user data'
      operationId: getUserV2
      parameters:
        -
          name: user
          in: path
          description: 'User id'
          required: true
          example: 1
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                properties:
                  user: { $ref: '#/components/schemas/UserResource' }
                type: object
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthorized
        '404':
          description: 'Not found'
      security:
        -
          sanctum: []
    put:
      tags:
        - Users
      summary: 'Update existing user'
      operationId: updateUserV2
      parameters:
        -
          name: user
          in: path
          description: 'User id'
          required: true
          example: 1
      requestBody:
        description: 'Update existing user'
        required: true
        content:
          application/json:
            schema:
              required:
                - first_name
                - last_name
                - email
                - password
              properties:
                first_name:
                  type: string
                  example: John
                last_name:
                  type: string
                  example: Doe
                email:
                  type: string
                  example: john.doe@q.agency
                password:
                  type: string
                  example: secretPassword
              type: object
      responses:
        '200':
          description: 'User updated successfully'
          content:
            application/json:
              schema:
                properties:
                  user: { $ref: '#/components/schemas/UserResource' }
                type: object
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthorized
        '422':
          description: 'Validation failed'
        '404':
          description: 'Not found'
      security:
        -
          sanctum: []
    delete:
      tags:
        - Users
      summary: 'Delete user'
      operationId: deleteUserV2
      parameters:
        -
          name: user
          in: path
          description: 'User id'
          required: true
          example: 1
      responses:
        '204':
          description: Deleted
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'User deleted.' }
                type: object
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthorized
        '404':
          description: 'Not found'
      security:
        -
          sanctum: []
  /api/v1/reset-password:
    post:
      tags:
        - Authentication
      summary: 'Send password reset link'
      description: 'Sends password reset link'
      operationId: forgotPassword
      requestBody:
        description: 'Send password reset link'
        required: true
        content:
          application/json:
            schema:
              required:
                - email
              properties:
                email:
                  type: string
                  example: john.doe@q.agency
              type: object
      responses:
        '201':
          description: 'Password reset link sent successfully'
        '400':
          description: 'Bad Request'
        '403':
          description: Forbidden
        '422':
          description: 'Validation failed'
  /api/v1/login:
    post:
      tags:
        - Authentication
      summary: 'Authenticate user'
      description: 'Authenticate existing user'
      operationId: loginUser
      requestBody:
        description: 'Login existing user'
        required: true
        content:
          application/json:
            schema:
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  example: john.doe@q.agency
                password:
                  type: string
                  example: secretPassword
              type: object
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                properties:
                  token: { type: string, example: 1|HM3l3TpqlOlOxMwMeFlzaVorpgyR0UFJzlPTHHdA }
                  token_type: { type: string, example: Bearer }
                  user: { $ref: '#/components/schemas/UserResource' }
                type: object
...
  '/api/v1/reset-password/{token}':
    put:
      tags:
        - Authentication
      summary: 'Update password'
      description: 'Update password'
      operationId: resetPassword
      parameters:
        -
          name: token
          in: path
          description: 'Token string'
          required: true
          schema:
            type: string
            example: 1|HM3l3TpqlOlOxMwMeFlzaVorpgyR0UFJzlPTHHdA
      requestBody:
        description: 'Update password'
        required: true
        content:
          application/json:
            schema:
              required:
                - email
                - password
                - password_confirmation
              properties:
                email:
                  type: string
                  example: john.doe@q.agency
                password:
                  type: string
                  example: secretPassword
                password_confirmation:
                  type: string
                  example: secretPassword
              type: object
      responses:
        '201':
          description: 'User created successfully'
        '400':
          description: 'Bad Request'
        '403':
          description: Forbidden
        '422':
          description: 'Validation failed'
components:
  schemas:
...
    UserResource:
      description: 'Class UserResource'
      properties:
        id:
          type: integer
          default: 1
        first_name:
          type: string
          default: John
        last_name:
          type: string
          default: Doe
        email:
          type: string
          default: john.doe@q.agency
      type: object
...